refactor: remove lib_standard_app files
- removed `src/common` files, now part of the ledger sdks
- removed `src/io.h/c` files,  now part of the ledger sdks
- removed `parser.h/c` files,  now part of the ledger sdks
- main.c => app_main.c, updated patterns
- refactored `handler_get_app_name` to match updated `app-boilerplate` pattern
- refactored `handler_get_version` to match updated `app-boilerplate` pattern
- custom `format.h/c` methods split into new files (ui_utils.h/c and transaction_utils.h/c)
