--- src/common/format.h
@@ -1,134 +1,0 @@
-/*****************************************************************************
- *  This work is licensed under a Creative Commons Attribution-NoDerivatives
- *  4.0 International License.
- *
- *  This software also incorporates work covered by the following copyright
- *  and permission notice:
- *
- *   (c) 2020 Ledger SAS.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *****************************************************************************/
-
-#pragma once
-
-#include <stddef.h>   // size_t
-#include <stdint.h>   // int*_t, uint*_t
-#include <stdbool.h>  // bool
-
-/**
- * Format 64-bit signed integer as string.
- *
- * @param[out] dst
- *   Pointer to output string.
- * @param[in]  dst_len
- *   Length of output string.
- * @param[in]  value
- *   64-bit signed integer to format.
- *
- * @return true if success, false otherwise.
- *
- */
-bool format_i64(char *dst, size_t dst_len, const int64_t value);
-
-/**
- * Format 64-bit unsigned integer as string.
- *
- * @param[out] dst
- *   Pointer to output string.
- * @param[in]  dst_len
- *   Length of output string.
- * @param[in]  value
- *   64-bit unsigned integer to format.
- *
- * @return true if success, false otherwise.
- *
- */
-bool format_u64(char *dst, size_t dst_len, uint64_t value);
-
-/**
- * Format 64-bit unsigned integer as string with decimals.
- *
- * @param[out] dst
- *   Pointer to output string.
- * @param[in]  dst_len
- *   Length of output string.
- * @param[in]  value
- *   64-bit unsigned integer to format.
- * @param[in]  decimals
- *   Number of digits after decimal separator.
- *
- * @return true if success, false otherwise.
- *
- */
-bool format_fpu64(char *dst, size_t dst_len, const uint64_t value, uint8_t decimals);
-
-/**
- * Format 64-bit unsigned integer as string with decimals and ticker.
- *
- * @param[out] dst
- *   Pointer to output string.
- * @param[in]  dst_len
- *   Length of output string.
- * @param[in]  value
- *   64-bit unsigned integer to format.
- * @param[in]  decimals
- *   Number of digits after decimal separator.
- * @param[in]  ticker
- *   The token's ticker name to be displayed.
- * @param[in]  ticker_len
- *   The length of the ticker name (not including the null-terminator).
- *
- * @return true if success, false otherwise.
- *
- */
-bool format_amount(char *dst,
-                   size_t dst_len,
-                   const uint64_t value,
-                   uint8_t decimals,
-                   const char *ticker,
-                   size_t ticker_len);
-
-/**
- * Format 16-bit unsigned integer as string with decimals and percentage sign.
- *
- * @param[out] dst
- *   Pointer to output string.
- * @param[in]  dst_len
- *   Length of output string.
- * @param[in]  value
- *   16-bit unsigned integer to format.
- * @param[in]  decimals
- *   Number of digits after decimal separator.
- *
- * @return true if success, false otherwise.
- *
- */
-bool format_percentage(char *dst, size_t dst_len, const uint16_t value, uint8_t decimals);
-
-/**
- * Format byte buffer to uppercase hexadecimal string.
- *
- * @param[in]  in
- *   Pointer to input byte buffer.
- * @param[in]  in_len
- *   Length of input byte buffer.
- * @param[out] out
- *   Pointer to output string.
- * @param[in]  out_len
- *   Length of output string.
- *
- * @return number of bytes written if success, -1 otherwise.
- *
- */
-int format_hex(const uint8_t *in, size_t in_len, char *out, size_t out_len);
