--- src/common/macros.h
@@ -1,6 +1,0 @@
-#pragma once
-
-/**
- * Macro for the size of a specific structure field.
- */
-#define MEMBER_SIZE(type, member) (sizeof(((type *) 0)->member))
