+++ src/handler/get_app_name.c
@@ -31,9 +31,7 @@
 #include "common/buffer.h"
 
 int handler_get_app_name() {
-    _Static_assert(APPNAME_LEN < MAX_APPNAME_LEN, "APPNAME must be at most 64 characters!");
+    _Static_assert(APPNAME_LEN < APPNAME_MAX_LEN, "APPNAME must be at most 64 characters!");
 
-    buffer_t rdata = {.ptr = (uint8_t *) PIC(APPNAME), .size = APPNAME_LEN, .offset = 0};
-
-    return io_send_response(&rdata, SW_OK);
+    return io_send_response_pointer(PIC(APPNAME), APPNAME_LEN, SW_OK);
 }
