+++ src/transaction/transaction_utils.h
@@ -6,7 +6,7 @@
  *  and permission notice:
  *
  *   Ledger App Boilerplate.
- *   (c) 2020 Ledger SAS.
+ *   (c) 2023 Ledger SAS.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -23,10 +23,52 @@
 
 #pragma once
 
+#include <stddef.h>   // size_t
 #include <stdint.h>   // uint*_t
 #include <stdbool.h>  // bool
 
-#include "types.h"
+/**
+ * Format 64-bit unsigned integer as string with decimals and ticker.
+ *
+ * @param[out] dst
+ *   Pointer to output string.
+ * @param[in]  dst_len
+ *   Length of output string.
+ * @param[in]  value
+ *   64-bit unsigned integer to format.
+ * @param[in]  decimals
+ *   Number of digits after decimal separator.
+ * @param[in]  ticker
+ *   The token's ticker name to be displayed.
+ * @param[in]  ticker_len
+ *   The length of the ticker name (not including the null-terminator).
+ *
+ * @return true if success, false otherwise.
+ *
+ */
+bool format_amount(char *dst,
+                   size_t dst_len,
+                   const uint64_t value,
+                   uint8_t decimals,
+                   const char *ticker,
+                   size_t ticker_len);
+
+/**
+ * Format 16-bit unsigned integer as string with decimals and percentage sign.
+ *
+ * @param[out] dst
+ *   Pointer to output string.
+ * @param[in]  dst_len
+ *   Length of output string.
+ * @param[in]  value
+ *   16-bit unsigned integer to format.
+ * @param[in]  decimals
+ *   Number of digits after decimal separator.
+ *
+ * @return true if success, false otherwise.
+ *
+ */
+bool format_percentage(char *dst, size_t dst_len, const uint16_t value, uint8_t decimals);
 
 /**
  * Check if input is encoded using ASCII characters.
@@ -34,7 +76,7 @@
  * @param[in] text
  *   Pointer to input byte buffer.
  * @param[in] text_len
- *   Lenght of input byte buffer.
+ *   Length of input byte buffer.
  * @param[in] allow_new_lines
  *   Allow ASCII text to contain new lines.
  *
